Resetting to latest V1 code

+++ RazorDBTests/CrashTests.cs
@@ -1,17 +1,18 @@
-﻿/* 
-Copyright 2012 Gnoso Inc.
+﻿/*
+Copyright 2012, 2013 Gnoso Inc.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+This software is licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except for what is in compliance with the License.
+
+You may obtain a copy of this license at
 
     http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
+
+See the License for the specific language governing permissions and limitations.
 */
 using System;
 using System.Collections.Generic;
@@ -57,7 +58,6 @@ namespace RazorDBTests {
                 int ct = 0;
                 foreach (var pair in db.Enumerate()) {
                     ByteArray k = new ByteArray(pair.Key);
-                    ByteArray v = new ByteArray(pair.Value);
                     Assert.True(lastKey.CompareTo(k) < 0);
                     lastKey = k;
                     ct++;
@@ -83,8 +83,10 @@ namespace RazorDBTests {
             var process = Process.Start(testPath, "CrashTestBeforeMerge");
 
             doneSetting.WaitOne(30000);
-            process.Kill();
-            process.WaitForExit();
+            if (!process.HasExited) {
+                process.Kill();
+                process.WaitForExit();
+            }
 
             // Open the database created by the other program
             using (var db = new KeyValueStore(path)) {
@@ -96,7 +98,6 @@ namespace RazorDBTests {
                 int ct = 0;
                 foreach (var pair in db.Enumerate()) {
                     ByteArray k = new ByteArray(pair.Key);
-                    ByteArray v = new ByteArray(pair.Value);
                     Assert.True(lastKey.CompareTo(k) < 0);
                     lastKey = k;
                     ct++;
